refactor: update PopularMediaUiState and simplify HomeViewModel
