+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/home/HomeScreen.kt
@@ -3,16 +3,12 @@ package com.berlin.aflami.screens.home
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.PageSize
-import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -22,7 +18,6 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
@@ -30,16 +25,14 @@ import androidx.compose.ui.unit.dp
 import com.berlin.aflami.component.BlurredPosterBackground
 import com.berlin.aflami.component.GenersChip
 import com.berlin.aflami.component.HomeBar
-import com.berlin.aflami.component.MovieCard
 import com.berlin.aflami.component.SectionTitle
 import com.berlin.aflami.screens.home.component.HomeSections
-import com.berlin.aflami.screens.home.screen.MoviesPosterSlider
+import com.berlin.aflami.screens.home.component.PosterSlider
 import com.berlin.aflami.ui.theme.Theme
 import com.berlin.aflami.viewmodel.home.HomeInteractionListener
 import com.berlin.aflami.viewmodel.home.HomeScreenEffect
 import com.berlin.aflami.viewmodel.home.HomeViewModel
 import com.berlin.aflami.viewmodel.home.uistate.HomeUiState
-import com.berlin.aflami.viewmodel.shareduistate.MediaUiState
 import com.berlin.ui.R
 import org.koin.androidx.compose.koinViewModel
 
@@ -74,9 +67,9 @@ fun HomeContent(
     Column (
         modifier = Modifier.padding(top = 6.dp)
     ){
-        val pagerState = rememberPagerState(initialPage = 0, pageCount = { mediaList.size })
+        val pagerState = rememberPagerState(initialPage = 0, pageCount = { state.popularMedia.popularMedia.size })
 
-        val currentMedia = mediaList.getOrNull(pagerState.currentPage)
+        val currentMedia = state.popularMedia.popularMedia.getOrNull(pagerState.currentPage)
 
         Column(
             modifier = Modifier
@@ -118,9 +111,9 @@ fun HomeContent(
                         }
                     )
 
-                    MoviesPosterSlider(
+                    PosterSlider(
                         modifier = Modifier.padding(top = 12.dp, bottom = 8.dp),
-                        mediaList = mediaList,
+                        mediaList = state.popularMedia.popularMedia,
                         pagerState = pagerState,
                         onClick = { }
                     )
@@ -158,39 +151,5 @@ fun HomeContent(
             state = state.mediaContinueWatching,
             sectionTitleId = R.string.continue_watching
         )
-    }
-
-
-
-
-
-}
-
-@Composable
-fun MoviesPosterSlider(
-    modifier: Modifier = Modifier,
-    mediaList: List<MediaUiState>,
-    onClick: (MediaUiState) -> Unit = {},
-    pagerState: PagerState,
-) {
-
-    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
-    val itemWidth = 244.dp
-    val contentPadding = (screenWidth - itemWidth) / 2
-    HorizontalPager(
-        state = pagerState,
-        pageSize = PageSize.Fixed(itemWidth),
-        contentPadding = PaddingValues(horizontal = contentPadding),
-        modifier = modifier.fillMaxWidth()
-    ) { pageIndex ->
-        val mediaItem = mediaList.getOrNull(pageIndex)
-        mediaItem?.let {
-            MovieCard(
-                isCentered = pageIndex == pagerState.currentPage,
-                onClick = { onClick(it) },
-                rating = it.rating,
-                posterImageUrl = it.poster
-            )
-        }
     }
 }
