+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/home/uistate/HomeUiState.kt
@@ -28,8 +28,6 @@ data class UpComingUiState(
 data class PopularMediaUiState (
     val isLoading: Boolean = false,
     val popularMedia: List<MediaUiState> = emptyList(),
-    val moviesOnly: List<MediaUiState> = emptyList(),
-    val tvShowsOnly: List<MediaUiState> = emptyList(),
     val error: String? = null
 )
